openconnect: fix do_compile failure
This patch solves the following error which appears now and then because of
parallel building.

  ../git/main.c:78:21: fatal error: version.c: No such file or directory

The problem is that version.c needs to be there when building some objects
which needs main.c. But version.c depends on main.c and is generated at
build time.

This patch fix the problem by generating version.c first before building others.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>